refactor text
